ARROW-11899: [Java] Refactor the compression codec implementation into core/Arrow specific parts
